requirements/dev.txt
httpx==0.24.1
fastapi==0.103.1
uvicorn==0.23.0
prometheus-client==0.17.0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
schedule==1.2.1